The big bad reformat

+++ src/web/core/src/test/java/org/geoserver/web/wicket/browser/FileDataViewTest.java
@@ -5,10 +5,10 @@
  */
 package org.geoserver.web.wicket.browser;
 
+import static org.geoserver.web.GeoServerWicketTestSupport.initResourceSettings;
+import static org.junit.Assert.*;
+
 import java.io.File;
-
-import junit.framework.TestCase;
-
 import org.apache.commons.io.FileUtils;
 import org.apache.wicket.Component;
 import org.apache.wicket.ajax.AjaxRequestTarget;
@@ -20,10 +20,7 @@ import org.geoserver.web.FormTestPage;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.geoserver.web.GeoServerWicketTestSupport.initResourceSettings;
-import static org.junit.Assert.*;
-
-public class FileDataViewTest{
+public class FileDataViewTest {
 
     private WicketTester tester;
 
@@ -32,40 +29,42 @@ public class FileDataViewTest{
     private File one;
 
     private File two;
-    
+
     private File lastClicked;
 
     FileProvider fileProvider;
-    
+
     @Before
     public void setUp() throws Exception {
         tester = new WicketTester();
         initResourceSettings(tester);
-        
+
         root = new File("target/test-dataview");
-        if(root.exists())
-            FileUtils.deleteDirectory(root);
+        if (root.exists()) FileUtils.deleteDirectory(root);
         root.mkdirs();
         one = new File(root, "one.txt");
         one.createNewFile();
         two = new File(root, "two.sld");
         two.createNewFile();
-        
+
         fileProvider = new FileProvider(root);
-        
-        tester.startPage(new FormTestPage(new ComponentBuilder() {
 
-            public Component buildComponent(String id) {
-                
-                return new FileDataView(id, fileProvider) {
+        tester.startPage(
+                new FormTestPage(
+                        new ComponentBuilder() {
 
-                    @Override
-                    protected void linkNameClicked(File file, AjaxRequestTarget target) {
-                        lastClicked = file;
-                    }
-                };
-            }
-        }));
+                            public Component buildComponent(String id) {
+
+                                return new FileDataView(id, fileProvider) {
+
+                                    @Override
+                                    protected void linkNameClicked(
+                                            File file, AjaxRequestTarget target) {
+                                        lastClicked = file;
+                                    }
+                                };
+                            }
+                        }));
 
         // WicketHierarchyPrinter.print(tester.getLastRenderedPage(), true, true);
     }
@@ -74,10 +73,15 @@ public class FileDataViewTest{
     public void testLoad() throws Exception {
         tester.assertRenderedPage(FormTestPage.class);
         tester.assertNoErrorMessage();
-        
+
         tester.assertLabel("form:panel:fileTable:fileContent:files:1:nameLink:name", "one.txt");
         tester.assertLabel("form:panel:fileTable:fileContent:files:2:nameLink:name", "two.sld");
-        assertEquals(2, ((DataView) tester.getComponentFromLastRenderedPage("form:panel:fileTable:fileContent:files")).size());
+        assertEquals(
+                2,
+                ((DataView)
+                                tester.getComponentFromLastRenderedPage(
+                                        "form:panel:fileTable:fileContent:files"))
+                        .size());
     }
 
     @Test
@@ -93,17 +97,21 @@ public class FileDataViewTest{
         fileProvider.setFileFilter(new Model(new ExtensionFileFilter(".txt")));
         tester.startPage(tester.getLastRenderedPage());
         tester.assertLabel("form:panel:fileTable:fileContent:files:3:nameLink:name", "one.txt");
-        assertEquals(1, ((DataView) tester.getComponentFromLastRenderedPage("form:panel:fileTable:fileContent:files")).size());
+        assertEquals(
+                1,
+                ((DataView)
+                                tester.getComponentFromLastRenderedPage(
+                                        "form:panel:fileTable:fileContent:files"))
+                        .size());
     }
-    
+
     public void testSortByName() throws Exception {
-        
-        
+
         // order by inverse name
         tester.clickLink("form:panel:fileTable:nameHeader:orderByLink", true);
         tester.clickLink("form:panel:fileTable:nameHeader:orderByLink", true);
         tester.assertRenderedPage(FormTestPage.class);
-         
+
         tester.assertLabel("form:panel:fileTable:fileContent:files:5:nameLink:name", "two.sld");
         tester.assertLabel("form:panel:fileTable:fileContent:files:6:nameLink:name", "one.txt");
     }
