+++ src/web/core/src/test/java/org/geoserver/web/wicket/browser/GeoServerFileChooserTest.java
@@ -9,7 +9,6 @@ import static org.junit.Assert.*;
 
 import java.io.File;
 import java.io.IOException;
-
 import org.apache.commons.io.FileUtils;
 import org.apache.wicket.Component;
 import org.apache.wicket.ajax.AjaxRequestHandler;
@@ -34,10 +33,9 @@ public class GeoServerFileChooserTest extends GeoServerWicketTestSupport {
 
     @Before
     public void init() throws IOException {
-        
+
         root = new File("target/test-filechooser");
-        if(root.exists())
-            FileUtils.deleteDirectory(root);
+        if (root.exists()) FileUtils.deleteDirectory(root);
         child = new File(root, "child");
         child.mkdirs();
         one = new File(child, "one.txt");
@@ -45,29 +43,37 @@ public class GeoServerFileChooserTest extends GeoServerWicketTestSupport {
         two = new File(child, "two.sld");
         two.createNewFile();
     }
-    
+
     public void setupChooser(final File file) {
-        tester.startPage(new FormTestPage(new ComponentBuilder() {
-            
-            public Component buildComponent(String id) {
-                return new GeoServerFileChooser(id, new Model(file));
-            }
-        }));
-        
-        //WicketHierarchyPrinter.print(tester.getLastRenderedPage(), true, true);
+        tester.startPage(
+                new FormTestPage(
+                        new ComponentBuilder() {
+
+                            public Component buildComponent(String id) {
+                                return new GeoServerFileChooser(id, new Model(file));
+                            }
+                        }));
+
+        // WicketHierarchyPrinter.print(tester.getLastRenderedPage(), true, true);
     }
-    
+
     @Test
     public void testLoad() {
         setupChooser(root);
-        
+
         tester.assertRenderedPage(FormTestPage.class);
         tester.assertNoErrorMessage();
-        
-        tester.assertLabel("form:panel:fileTable:fileTable:fileContent:files:1:nameLink:name", "child/");
-        assertEquals(1, ((DataView) tester.getComponentFromLastRenderedPage("form:panel:fileTable:fileTable:fileContent:files")).size());
+
+        tester.assertLabel(
+                "form:panel:fileTable:fileTable:fileContent:files:1:nameLink:name", "child/");
+        assertEquals(
+                1,
+                ((DataView)
+                                tester.getComponentFromLastRenderedPage(
+                                        "form:panel:fileTable:fileTable:fileContent:files"))
+                        .size());
     }
-    
+
     @Test
     public void testNullRoot() {
         setupChooser(null);
@@ -75,59 +81,66 @@ public class GeoServerFileChooserTest extends GeoServerWicketTestSupport {
         // make sure it does not now blow out because of the null
         tester.assertRenderedPage(FormTestPage.class);
         tester.assertNoErrorMessage();
-        
-        tester.assertLabel("form:panel:breadcrumbs:path:0:pathItemLink:pathItem", getTestData().getDataDirectoryRoot().getName() + "/");
+
+        tester.assertLabel(
+                "form:panel:breadcrumbs:path:0:pathItemLink:pathItem",
+                getTestData().getDataDirectoryRoot().getName() + "/");
     }
-    
+
     @Test
     public void testInDialog() throws Exception {
-        tester.startPage(new FormTestPage(new ComponentBuilder() {
-            
-            public Component buildComponent(String id) {
-                return new GeoServerDialog(id);
-            }
-        }));
-        
+        tester.startPage(
+                new FormTestPage(
+                        new ComponentBuilder() {
+
+                            public Component buildComponent(String id) {
+                                return new GeoServerDialog(id);
+                            }
+                        }));
+
         tester.assertRenderedPage(FormTestPage.class);
 
         tester.debugComponentTrees();
-        
-        GeoServerDialog dialog = (GeoServerDialog) tester.getComponentFromLastRenderedPage("form:panel");
+
+        GeoServerDialog dialog =
+                (GeoServerDialog) tester.getComponentFromLastRenderedPage("form:panel");
         assertNotNull(dialog);
 
-        dialog.showOkCancel(new AjaxRequestHandler(tester.getLastRenderedPage()),
-            new DialogDelegate() {
-                @Override
-                protected Component getContents(String id) {
-                    return new GeoServerFileChooser(id, new Model(root));
-                }
+        dialog.showOkCancel(
+                new AjaxRequestHandler(tester.getLastRenderedPage()),
+                new DialogDelegate() {
+                    @Override
+                    protected Component getContents(String id) {
+                        return new GeoServerFileChooser(id, new Model(root));
+                    }
 
-                @Override
-                protected boolean onSubmit(AjaxRequestTarget target, Component contents) {
-                    assertNotNull(contents);
-                    assertTrue(contents instanceof GeoServerFileChooser);
-                    return true;
-                }
-        });
-        
+                    @Override
+                    protected boolean onSubmit(AjaxRequestTarget target, Component contents) {
+                        assertNotNull(contents);
+                        assertTrue(contents instanceof GeoServerFileChooser);
+                        return true;
+                    }
+                });
+
         dialog.submit(new AjaxRequestHandler(tester.getLastRenderedPage()));
     }
 
     @Test
     public void testHideFileSystem() throws Exception {
-        tester.startPage(new FormTestPage(new ComponentBuilder() {
-            public Component buildComponent(String id) {
-                return new GeoServerFileChooser(id, new Model(), true);
-            }
-        }));
-        
+        tester.startPage(
+                new FormTestPage(
+                        new ComponentBuilder() {
+                            public Component buildComponent(String id) {
+                                return new GeoServerFileChooser(id, new Model(), true);
+                            }
+                        }));
+
         tester.assertRenderedPage(FormTestPage.class);
         tester.assertNoErrorMessage();
-        
-        DropDownChoice<File> rootChoice = 
-            (DropDownChoice<File>) tester.getComponentFromLastRenderedPage("form:panel:roots");
+
+        DropDownChoice<File> rootChoice =
+                (DropDownChoice<File>) tester.getComponentFromLastRenderedPage("form:panel:roots");
         assertEquals(1, rootChoice.getChoices().size());
         assertEquals(getDataDirectory().root(), rootChoice.getChoices().get(0));
-        
     }
 }
